fixed some code after code review
