+++ .expo/packager-info.json
@@ -1,8 +1,8 @@
 {
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 26350,
+  "packagerPid": 34622,
   "expoServerNgrokUrl": "https://e6-f5f.iningcha.eventall.exp.direct",
   "packagerNgrokUrl": "https://packager.e6-f5f.iningcha.eventall.exp.direct",
-  "ngrokPid": 26389
+  "ngrokPid": 34643
 }
