+++ src/EventPage/screens/EventPageContainer.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import EventPage from './EventPage';
+
+class EventPageContainer extends React.Component {
+  state = {
+    event: { title: '' },
+  }
+  componentDidMount() {
+    this.setState({ event: { title: 'Creating Reality Hackthon', map: 'picture' } });
+  }
+
+  render() {
+    return <EventPage event={this.state.event} />;
+  }
+}
+
+export default EventPageContainer;
+
