+++ src/globals/colors.js
@@ -1,3 +1,4 @@
 export default {
   white: '#ffffff',
+  backgroundgrey: '#F2F2F2',
 };
